.TinyMDE full height
Make the editor area be full height of the container element, so that clicking anywhere in the container brings focus to the editor. Just the first thing I did after incorporating TinyMDE in my project, and I thought maybe it's a nice default to have?

Thanks for a great library!